Refactored local and dask to deduplicate wrapper code
local and dask wrappers are now stored in a separate module under
covalent/executor/utils
