+++ covalent/executor/executor_plugins/dask.py
@@ -25,10 +25,7 @@ and waits for execution to finish then returns the result.
 This is a plugin executor module; it is loaded if found and properly structured.
 """
 
-import io
 import os
-import traceback
-from contextlib import redirect_stderr, redirect_stdout
 from typing import Callable, Dict, List
 
 from dask.distributed import Client
@@ -39,6 +36,7 @@ from covalent._shared_files import TaskRuntimeError, logger
 from covalent._shared_files.config import get_config
 from covalent._shared_files.utils import _address_client_mapper
 from covalent.executor.base import AsyncBaseExecutor
+from covalent.executor.utils.wrappers import io_wrapper as dask_wrapper
 
 # The plugin class name must be given by the executor_plugin_name attribute:
 EXECUTOR_PLUGIN_NAME = "DaskExecutor"
@@ -53,18 +51,6 @@ _EXECUTOR_PLUGIN_DEFAULTS = {
         os.environ.get("XDG_CACHE_HOME") or os.path.join(os.environ["HOME"], ".cache"), "covalent"
     ),
 }
-
-
-def dask_wrapper(fn, args, kwargs):
-    with redirect_stdout(io.StringIO()) as stdout, redirect_stderr(io.StringIO()) as stderr:
-        try:
-            output = fn(*args, **kwargs)
-            tb = ""
-        except Exception as ex:
-            output = None
-            tb = "".join(traceback.TracebackException.from_exception(ex).format())
-
-    return output, stdout.getvalue(), stderr.getvalue(), tb
 
 
 class DaskExecutor(AsyncBaseExecutor):
