+++ covalent/executor/utils/wrappers.py
@@ -26,12 +26,12 @@ import io
 import multiprocessing as mp
 import traceback
 from contextlib import redirect_stderr, redirect_stdout
-from typing import Callable, Dict, List
+from typing import Any, Callable, Dict, List, Tuple
 
 
-def local_wrapper(fn: Callable, args: List, kwargs: Dict, q: mp.Queue):
+def io_wrapper(fn: Callable, args: List, kwargs: Dict) -> Tuple[Any, str, str, str]:
     """Wrapper function to execute the given function in a separate
-    process and pass the output back to the runner"""
+    process and capture stdout and stderr"""
     with redirect_stdout(io.StringIO()) as stdout, redirect_stderr(io.StringIO()) as stderr:
         try:
             output = fn(*args, **kwargs)
@@ -39,5 +39,17 @@ def local_wrapper(fn: Callable, args: List, kwargs: Dict, q: mp.Queue):
         except Exception as ex:
             output = None
             tb = "".join(traceback.TracebackException.from_exception(ex).format())
+    return output, stdout.getvalue(), stderr.getvalue(), tb
 
-    q.put((output, stdout.getvalue(), stderr.getvalue(), tb))
+
+def local_wrapper(fn: Callable, args: List, kwargs: Dict, q: mp.Queue):
+    """Wrapper function for use with LocalExecutor
+
+    Sends function output back to the main dispatcher process through
+    a multiprocessing queue
+
+    """
+
+    output, stdout, stderr, tb = io_wrapper(fn, args, kwargs)
+
+    q.put((output, stdout, stderr, tb))
